+++ src/types/types.ts
@@ -600,6 +600,15 @@ export interface SpamInfo {
   classifications: NftSpamClassification[];
 }
 
+/** Block time and number at which an NFT was acquired. */
+export interface AcquiredAt {
+  /** Timestamp of the block at which an NFT was last acquired. */
+  blockTimestamp?: string;
+
+  /** Block number of the block at which an NFT was last acquired. */
+  blockNumber?: number;
+}
+
 /**
  * Optional parameters object for the {@link getNftsForOwner} and
  * {@link getNftsForOwnerIterator} functions.
@@ -1317,8 +1326,16 @@ export interface NftSaleFeeData {
   /** The symbol of the token used for the payment. */
   symbol: string;
 
-  /** The number of decimals of the token used for the payment. */
+  /**
+   * The number of decimals of the token used for the payment.
+   * @deprecated Please use `decimals` instead. This field contains a typo
+   * and will always be undefined. It's been marked as deprecated to avoid
+   * breaking changes.
+   */
   decimal: number;
+
+  /** The number of decimals of the token used for the payment. */
+  decimals: number;
 }
 
 /**