+++ test/test-util.ts
@@ -209,6 +209,7 @@ export function createRawNftSale(
   const feeData: NftSaleFeeData = {
     amount: '100',
     decimal: 18,
+    decimals: 18,
     symbol: 'ETH'
   };